+++ crates/warpgate/src/loader.rs
@@ -127,8 +127,12 @@ impl PluginLoader {
             sha.update(seed);
         }
 
+        // Remove unwanted or unsafe file name characters
+        let safe_id = id.as_str().replace(['/', '@', '.', ' '], "");
+
         self.plugins_dir.join(format!(
-            "{id}{}{:x}{}",
+            "{}{}{:x}{}",
+            safe_id,
             if is_latest { "-latest-" } else { "-" },
             sha.finalize(),
             determine_cache_extension(url)